requirements.txt
# PDF 처리
pdfplumber>=0.9.0

# OpenAI
openai>=1.3.0
python-dotenv>=1.0.0

# Vector DB
chromadb>=0.4.0
pinecone>=3.0.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langgraph>=0.0.20

# 유틸리티
numpy>=1.24.0
pandas>=2.0.0
tiktoken>=0.5.0

# 웹 인터페이스
streamlit>=1.28.0
